split createServer in 4 plugins
