+++ src/version.js
@@ -1,6 +1,6 @@
 'use strict';
 
 module.exports={
-  defaultVersion:'1.8',
+  defaultVersion:'1.12',
   supportedVersions:['1.7','1.8','1.9','1.10','1.11.2','1.12']
 };
